requirements.txt
PyQt5==5.15.9
requests==2.31.0
yt-dlp==2023.11.16
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
dropbox==11.36.2
tqdm>=4.62.3
beautifulsoup4>=4.12.0
schedule>=1.2.0
cryptography>=41.0.0
python-magic>=0.4.27
apscheduler>=3.10.0
psutil>=5.9.0
python-qbittorrent>=0.4.3
PySocks>=1.7.1
moviepy>=1.0.3
Pillow>=10.0.0
ffmpeg-python>=0.2.0
